feat: Large push
Finished and working S3 backend.

Beginning of BitWarden which looks like it will require the cli if i can't find a better solution.

Serialization which stores the config on the remote device, Which means i can worry about encrypting the passwords and so on later.

Implemented Auto-Pruner, Allows complete disabling, keeping files for x amount of time, and keeping a minimum amount of files despite the possible expired age of them.